Restructure to support new vendor tests
