+++ pom.xml
@@ -67,7 +67,7 @@
         <mjson.version>1.4.0</mjson.version>
         <orientdb.version>3.2.1.1</orientdb.version>
         <logback.version>1.1.7</logback.version>
-        <commons-io.version>1.3.2</commons-io.version>
+        <junit.version>4.11</junit.version>
     </properties>
 
     <licenses>
